use Drupal placeholders for CSS and JS
