--- web/themes/new_weather_theme/assets/js/components/index.js
@@ -1,5 +1,0 @@
-import locationSearch from "./locationSearch.js";
-
-export default () => {
-  locationSearch();
-};
