+++ web/themes/new_weather_theme/new_weather_theme.libraries.yml
@@ -1,105 +1,31 @@
-base:
-  version: VERSION
+uswds-init:
+  version: 3
+  header: true
+  js:
+    assets/js/uswds-init.js:
+      minified: true
+
+uswds-customized:
+  version: 3
   css:
-    component:
-      css/components/action-links.css:
-        weight: -10
-      css/components/breadcrumb.css:
-        weight: -10
-      css/components/button.css:
-        weight: -10
-      css/components/container-inline.css:
-        weight: -10
-      css/components/details.css:
-        weight: -10
-      css/components/exposed-filters.css:
-        weight: -10
-      css/components/field.css:
-        weight: -10
-      css/components/form.css:
-        weight: -10
-      css/components/icons.css:
-        weight: -10
-      css/components/inline-form.css:
-        weight: -10
-      css/components/item-list.css:
-        weight: -10
-      css/components/link.css:
-        weight: -10
-      css/components/links.css:
-        weight: -10
-      css/components/menu.css:
-        weight: -10
-      css/components/more-link.css:
-        weight: -10
-      css/components/pager.css:
-        weight: -10
-      css/components/tabledrag.css:
-        weight: -10
-      css/components/tableselect.css:
-        weight: -10
-      css/components/tablesort.css:
-        weight: -10
-      css/components/tabs.css:
-        weight: -10
-      css/components/textarea.css:
-        weight: -10
-      css/components/ui-dialog.css:
-        weight: -10
-dialog:
-  version: VERSION
-  css:
-    component:
-      css/components/dialog.css:
-        weight: -10
-dropbutton:
-  version: VERSION
-  css:
-    component:
-      css/components/dropbutton.css:
-        weight: -10
-file:
-  version: VERSION
-  css:
-    component:
-      css/components/file.css:
-        weight: -10
-image-widget:
-  version: VERSION
-  css:
-    component:
-      css/components/image-widget.css: {  }
-indented:
-  version: VERSION
-  css:
-    component:
-      css/components/indented.css: {  }
-messages:
-  version: VERSION
-  css:
-    component:
-      css/components/messages.css:
-        weight: -10
-node:
-  version: VERSION
-  css:
-    component:
-      css/components/node.css:
-        weight: -10
-progress:
-  version: VERSION
-  css:
-    component:
-      css/components/progress.css:
-        weight: -10
-search-results:
-  version: VERSION
-  css:
-    component:
-      css/components/search-results.css: {  }
-user:
-  version: VERSION
-  css:
-    component:
-      css/components/user.css:
-        weight: -10
+    theme:
+      assets/css/styles.css: {}
+  js:
+    assets/js/uswds.js:
+      attributes:
+        async: true
+      minified: true
+
+location-search:
+  version: 1
+  js:
+    assets/js/locationSearch.js:
+      attributes:
+        async: true
+
+localize-timestamps:
+  version: 1
+  js:
+    assets/js/localizeTimestamps.js:
+      attributes:
+        async: true
