+++ web/themes/new_weather_theme/templates/block/block--weathergov-current-conditions.html.twig
@@ -1,3 +1,5 @@
+{{ attach_library("new_weather_theme/localize-timestamps") }}
+
 {# Widgets and stuff. This is presented as a row of columns. #}
 <div class="weather-gov-current-conditions">
   <h2>Current Conditions</h2>
