+++ web/themes/new_weather_theme/templates/block/block--weathergov-location-search.html.twig
@@ -1,5 +1,5 @@
+{{ attach_library("new_weather_theme/location-search") }}
 <div>
-
   {% if content.place %}
   <h1 class="margin-bottom-0">
     {{ content.place.city }}{% if content.place.state %}, {{ content.place.state }}{% endif %}
