+++ web/themes/new_weather_theme/templates/layout/html.html.twig
@@ -38,9 +38,7 @@
     <title>{{ head_title|safe_join(' | ') }}</title>
     <css-placeholder token="{{ placeholder_token }}">
     <js-placeholder token="{{ placeholder_token }}">
-    <link rel="stylesheet" href="/{{ directory }}/assets/css/styles.css" />
     <link rel="icon" type="image/png" href="/{{ directory }}/assets/images/weather/favicons/noaa-favicon.ico" />
-    <script type="text/javascript" src="/{{ directory }}/assets/js/uswds-init.min.js"></script>
   </head>
   <body{{ attributes.addClass(body_classes) }}>
     {#
@@ -54,8 +52,5 @@
     {{ page }}
     {{ page_bottom }}
     <js-bottom-placeholder token="{{ placeholder_token }}">
-     <script type="text/javascript" defer src="/{{ directory }}/assets/js/uswds.min.js"></script>
-     <script type="module" async src="/{{ directory }}/assets/js/main.js"
-    ></script>
   </body>
 </html>
